+++ src/main.js
@@ -382,6 +382,10 @@ let mainfunc = {
           this.exportSelfMsg(`root属性${n}设置冲突，请检查!`)
         } else {
           this[n] = root[n]
+          let type = getType(this[n])
+          if (type !== 'object') {
+            this.exportSelfMsg(`root属性${n}类型为${type}，非object的值推荐赋值到data对象中，否则无法构建为响应式数据！`, 'waring')
+          }
         }
       }
     }